+++ src/main/java/SKM.java
@@ -1,0 +1,8 @@
+import Interface.Interface;
+import javafx.application.Application;
+
+public class SKM {
+    public static void main(String[] args) {
+        Application.launch(Interface.class, args);
+    }
+}