+++ src/main/java/Utils/Db/DBTool.java
@@ -1,4 +1,4 @@
-package Utils;
+package Utils.Db;
 
 import java.sql.Connection;
 import java.sql.DriverManager;