+++ src/main/java/Utils/FileRw/FileWriter.java
@@ -1,4 +1,8 @@
-import Utils.L10N;
+package Utils.FileRw;
+
+import Interface.L10N;
+import Interface.ShowPrompt;
+import Model.Key;
 import javafx.collections.ObservableList;
 import javafx.stage.FileChooser;
 import javafx.stage.Stage;