+++ src/main/java/Utils/Net/WebParser.java
@@ -1,4 +1,4 @@
-package Utils;
+package Utils.Net;
 
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpResponse;